refactor(perl-parser): split incremental module by responsibility (#0000)
