+++ crates/perl-parser/src/incremental/strategy.rs
@@ -1,0 +1,1 @@
+pub const MAX_EDIT_SIZE: usize = 64 * 1024;
